refactor: BaseP2PLibp2pConnectionAEATestCaseMany
